3 Features finished:HighLowGame, TicTacToe, and DateFact
